Refactored removal method
Created a findNode method and getParent method in order to find nodes
and get their parent nodes to re-establish links to nodes after deleting
them. (i.e. when the node to be deleted does not have any children, we
still have to go back and set what was pointing to it to null by using
it's parent node)